requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
playwright==1.43.0
python-dotenv==1.0.1
supabase==1.0.3
